feat: ShortCut -> ProfileNavItem 변경
